Add simple VR.Left/Right helper class, Blitter class
The Blitter class is meant to centralize and wrap the blitting, as
well as holding RenderTextures and Materials.

I'm finding it much easier to know what's going on with the
masking pipeline now that it isn't strewn about ClipManager,
and it will probably be even starker when no longer strewn
about Outliner/SwatchPicker.