Add various tweaks, sketch of new difference masking shader
This is a snapshot of settings that work with a particularly hard
video to filter.

Unfortunately, it only seems to work because of a recursive "innerThresh"
that's used to dampen the inclusion of non-moving pixels, but it's
turned up so high here that feed-forward amplification causes it to
"eat" the figure in most other videos.

It might be helpful to either eliminate the feedback interference
by rendering that particular alpha map on a separate pass that doesn't
feed-foward to subsequent frames, OR to refashion it into a togglable
"aggressive" mode and make it adjustable.